Refactor to get nifti header from byte array in one function
