+++ python/niizarr/__init__.py
@@ -1,2 +1,3 @@
 from ._nii2zarr import nii2zarr, nii2json   # noqa: F401
-from ._zarr2nii import zarr2nii, bin2nii    # noqa: F401
+from ._zarr2nii import zarr2nii    # noqa: F401
+from ._header import bin2nii
